[81] renamed file enum
