+++ src/app/components/popup-English/popup-english.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { EnglishLevels } from '../../enums/enums';
+import { EnglishLevels } from '../../enums/english-levels.enum';
 
 @Component({
   selector: 'app-popup-english',
